refactor: strengthen integer type semantics for massive task scalability
This is not a bug fix, the existing scheduler already proved stable
across 100M-task runs. This change is a proactive type-level upgrade to
unlock reliable testing and execution well beyond the 2^31 barrier.

Core changes:
- Migrated all scheduler counters (`m_tasks_in_progress`,
  `m_tasks_completed`, `m_tasks_failed`, `m_fallback_pending`) from
  `int` / `size_t` to `int64_t`.
    - Prevents silent overflow past INT_MAX (2.147B tasks).
    - Exposes underflow bugs as negative values rather than wrapping
      to huge unsigned numbers, simplifying debugging.
- Removed all `long` casts from `ChaseLevDeque` and unified size
  calculations to `int64_t`. `long` is 32-bit on Windows, which made
  large index arithmetic unsafe; this ensures true cross-platform 64-bit
  semantics.
- Updated public API: `GetTasksQueued()`, `GetTasksStolen()`, and all
  stat getters now consistently return `int64_t`.
- Dashboard integration: baselines, deltas and renderer values now use
  `int64_t` to prevent truncation in long-running benchmarks. Input
  parsing switched to `std::stoll` and loop counters to `size_t`.
- Workers: `steal_count` is now an `std::atomic<int64_t>` to match the
  new scheduler contract.

The scheduler can now be stress-tested with 3B, 5B or moretasks without the stats wrapping or the UI lying about progress. The type layer is now fully portable across Windows, Linux and macOS, paving the way for upcoming high-scale benchmarks.

+++ FluX/Scheduler.hpp
@@ -15,6 +15,7 @@
 #include "Workers.hpp"
 #include "TaskRegistry.hpp"
 
+
 /// High-level task scheduler with work-stealing thread pool.
 /// 
 /// Public API for task submission, tracking, and control.
@@ -64,15 +65,15 @@ public:
     Task GetTaskById(uint64_t task_id);
 
     /// Wait-free statistics (atomic loads)
-    int GetTasksInProgress() const { return m_tasks_in_progress.load(); }
-    int GetTasksCompleted() const { return m_tasks_completed.load(); }
-    int GetTasksFailed() const { return m_tasks_failed.load(); }
+    int64_t GetTasksInProgress() const { return m_tasks_in_progress.load(); }
+    int64_t GetTasksCompleted() const { return m_tasks_completed.load(); }
+    int64_t GetTasksFailed() const { return m_tasks_failed.load(); }
     /// Wait-free (relaxed atomic load): no longer takes m_mutex. See
     /// m_fallback_pending below -- kept in sync with m_tasks_to_dispatch by
     /// every mutating call site (EnqueuFallback / DrainFallbackQueue /
     /// DeleteTaskFromQueue / Run), so it never needs the lock to read.
-    size_t GetTasksQueued() const { return m_fallback_pending.load(std::memory_order_relaxed); }
-    uint64_t GetTasksStolen() const;
+    int64_t GetTasksQueued() const { return m_fallback_pending.load(std::memory_order_relaxed); }
+    int64_t GetTasksStolen() const;
 
 private:
     int GetHardwareThreadsCount() const;
@@ -104,13 +105,13 @@ private:
     /// common case where the fallback queue is empty -- which after Run()
     /// starts is effectively always, since AddTask only falls back when
     /// Workers::SubmitTask() itself reports saturation.
-    std::atomic<size_t> m_fallback_pending{ 0 };
+    std::atomic<int64_t> m_fallback_pending{ 0 };
 
     TaskRegistry<> m_registry;                            ///< Task metadata by ID (lock-free hot path, see TaskRegistry.hpp).
 
-    std::atomic<int>      m_tasks_in_progress{ 0 };        ///< Tasks currently executing.
-    std::atomic<int>      m_tasks_completed{ 0 };          ///< Tasks finished successfully.
-    std::atomic<int>      m_tasks_failed{ 0 };             ///< Tasks failed/aborted.
+    std::atomic<int64_t>      m_tasks_in_progress{ 0 };        ///< Tasks currently executing.
+    std::atomic<int64_t>      m_tasks_completed{ 0 };          ///< Tasks finished successfully.
+    std::atomic<int64_t>      m_tasks_failed{ 0 };             ///< Tasks failed/aborted.
 
     std::atomic<uint64_t> m_next_task_id{ 1 };             ///< Monotonic task ID generator.
